Adds HelpText hint customization and extract out to libs
