+++ app/assets/stylesheets/dreamsis.css.erb
@@ -2811,6 +2811,12 @@ form.formtastic > fieldset > ol > li > label {
 	padding-right: 3%;
 }
 
+.formtastic .label .button.help-text {
+  display: inline-block;
+  float: none;
+  margin-left: 0.5em;
+}
+
 /*input.currency:before {
   content: "$ ";
   float: left;
