+++ app/models/help_text.rb
@@ -1,5 +1,5 @@
 class HelpText < ActiveRecord::Base
-  attr_accessible :attribute_name, :instructions, :object_class, :title
+  attr_accessible :attribute_name, :instructions, :object_class, :title, :hint
   validates_uniqueness_of :attribute_name, :scope => [:object_class]
   
   def self.for(object_class, attribute_name)
