+++ db/migrate/20150904055454_add_hint_to_help_text.rb
@@ -1,0 +1,5 @@
+class AddHintToHelpText < ActiveRecord::Migration
+  def change
+    add_column :help_texts, :hint, :text
+  end
+end
